hamu_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
idna==3.10
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
